+++ src/modules/general/components/Button/index.tsx
@@ -1,8 +1,8 @@
 import { Button as MaterialButton } from '@mui/material';
-import * as React from 'react';
 
-import css from './button.module.scss';
-import { ButtonProps } from './button.types';
+import css from './index.module.scss';
+import { ButtonProps } from './index.types';
+
 export const Button: React.FC<ButtonProps> = ({
   children,
   customStyle,
