+++ src/modules/general/components/Button/index.types.ts
@@ -1,9 +1,9 @@
-import { ButtonProps as DefaultProps, LinkProps } from '@mui/material';
+import { ButtonProps as DefaultProps } from '@mui/material';
 
 export interface ButtonProps extends DefaultProps {
   customStyle?: string;
   block?: boolean;
-  color: 'primary' | 'secondary' | 'error' | 'info' | 'inherit';
+  color: 'primary' | 'secondary' | 'error' | 'warning' | 'info' | 'inherit';
   to?: string;
   target?: '_blank' | '_top' | '_self' | '_parent';
 }
