+++ src/modules/general/containers/ConfirmModal/index.types.ts
@@ -1,4 +1,4 @@
-import { ButtonProps } from 'src/modules/general/components/Button/button.types';
+import { ButtonProps } from 'src/modules/general/components/Button/index.types';
 import { ModalProps } from 'src/modules/general/components/modal/modal.types';
 
 export interface ConfirmModalProps extends ModalProps {
