+++ src/modules/general/containers/EmptyBox/index.types.ts
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import { ButtonProps } from 'src/modules/general/components/Button/button.types';
+import { ButtonProps } from 'src/modules/general/components/Button/index.types';
 
 export interface EmptyBoxProps {
   icon?: ReactNode;
